SimSwap/requirements.txt
imageio==2.9.0
torch==1.8.0
torchvision==0.9.0
numpy==1.21.1
insightface==0.2.1
ipython==7.21.0
Pillow==8.3.1
opencv-python==4.5.3.56
Fraction==1.5.1
onnxruntime-gpu==1.8.1
moviepy==1.0.3
